+++ src/stores/RemarkStore.tsx
@@ -1,4 +1,5 @@
 import { createStore } from "solid-js/store";
+import { Auxiliary } from "./ResponseStore";
 
 type Comment = {
     sender: any
@@ -13,42 +14,18 @@ type Note = {
 
 type Detail = {
     dataKey: string
-    templateDataKey: string
-    gearVersion: string
-    templateVersion: string
-    validationVersion: string
-    createdBy: string
-    updatedBy: string
-    createdAt: any
-    updatedAt: any
-    createdAtTimezone: string
-    createdAtGMT: number
-    updatedAtTimezone: string
-    updatedAtGMT: number
     notes: Note[]
 }
   
 export interface Remark{
     status: number
-    details: Detail
+    details: Detail & Auxiliary
 }
 
 export const [remark, setRemark] = createStore<Remark>({
     status: 1,
     details: {
         dataKey: '',
-        templateDataKey: '',
-        gearVersion: '',
-        templateVersion: '',
-        validationVersion: '',
-        createdBy: '',
-        updatedBy: '',
-        createdAt: '',
-        updatedAt: '',
-        createdAtTimezone: '',
-        createdAtGMT: null,
-        updatedAtTimezone: '',
-        updatedAtGMT: null,
         notes: []
     }
 });