+++ src/stores/ResponseStore.tsx
@@ -1,53 +1,43 @@
 import { createStore } from "solid-js/store";
-import { Summary } from "./SummaryStore"
+import { Summary } from "./SummaryStore";
 
-type Answer = {
+export type Answer = {
     dataKey: string;
     answer: any;
 }
 
+export type Auxiliary = {    
+    templateDataKey?: string
+    gearVersion?: string
+    templateVersion?: string
+    validationVersion?: string
+    createdBy?: string
+    updatedBy?: string
+    createdAt?: any
+    updatedAt?: any
+    createdAtTimezone?: string
+    createdAtGMT?: number
+    updatedAtTimezone?: string
+    updatedAtGMT?: number
+}
+
 type Detail = {
-    description: string
     dataKey: string
-    templateDataKey: string
-    gearVersion: string
-    templateVersion: string
-    validationVersion: string
+    description?: string
     docState?: string
-    createdBy: string
-    updatedBy: string
-    createdAt: any
-    updatedAt: any
-    createdAtTimezone: string
-    createdAtGMT: number
-    updatedAtTimezone: string
-    updatedAtGMT: number
     answers: Answer[]
     summary: Summary[]
 }
   
 export interface Response{
     status: number,
-    details: Detail
+    details: Detail & Auxiliary
 }
 
 export const [response, setResponse] = createStore<Response>({
     status: 1,
     details: {
-        description: '',
         dataKey: '',
-        templateDataKey: '',
-        gearVersion: '',
-        templateVersion: '',
-        validationVersion: '',
-        createdBy: '',
-        updatedBy: '',
-        createdAt: '',
-        updatedAt: '',
-        createdAtTimezone: '',
-        createdAtGMT: null,
-        updatedAtTimezone: '',
-        updatedAtGMT: null,
         answers: [],
         summary: [],
     }